Move task files to lib
