+++ bin/pelias-geonames
@@ -4,6 +4,8 @@ var program = require('commander-plus'),
     countryInfo = require('../metadata/countryInfo'),
     logger = require( 'pelias-logger' ).get( 'geonames' );
 
+var resolvers = require( '../lib/tasks/resolvers' );
+
 program
   .version('0.0.0')
   .usage('[options]')
@@ -14,7 +16,7 @@ program
   .parse(process.argv);
 
 if( program.countrycodes ){
-  var task = require('../task/countrycodes');
+  var task = require('../lib/tasks/countrycodes');
   task();
 }
 
@@ -23,18 +25,18 @@ else if( program.import ){
   var filename = isocode === 'ALL' ? 'allCountries' : isocode;
   var source = resolvers.selectSource( filename );
 
-  var task = require('../task/import');
+  var task = require('../lib/tasks/import');
   task( source );
 }
 
 else if( program.download ){
   var isocode = validateISOCode( program.download );
-  var task = require('../task/download');
+  var task = require('../lib/tasks/download');
   task( isocode === 'ALL' ? 'allCountries' : isocode );
 }
 
 else if( program.meta ){
-  var task = require('../task/meta');
+  var task = require('../lib/tasks/meta');
   task();
 }
