+++ lib/tasks/import.js
@@ -4,8 +4,8 @@ var peliasAdminLookup = require( 'pelias-admin-lookup' );
 var dbclient = require('pelias-dbclient');
 var model = require( 'pelias-model' );
 
-var adminLookupMetaStream = require('../lib/streams/adminLookupMetaStream');
-var peliasDocGenerator = require( '../lib/streams/peliasDocGenerator');
+var adminLookupMetaStream = require('../streams/adminLookupMetaStream');
+var peliasDocGenerator = require( '../streams/peliasDocGenerator');
 
 module.exports = function( sourceStream, endStream, config ){
   endStream = endStream || dbclient();
